requirements.txt
# Core
python-dotenv==1.0.0

# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
streamlit==1.31.0

# AI & NLP
anthropic==0.18.0
openai==1.12.0
spacy==3.7.2
sentence-transformers==2.3.1

# Vector DB
chromadb==0.4.22

# Data Processing
pandas==2.2.0
numpy==1.26.3
beautifulsoup4==4.12.3
requests==2.31.0
pypdf==4.2.0
pdf2image==1.17.0
pytesseract==0.3.10
Pillow==10.2.0
python-docx==1.1.2
openpyxl==3.1.5
fpdf2==2.7.9

# Visualization
plotly==5.18.0
networkx==3.2.1

# Database
sqlalchemy==2.0.25

# Utils
pydantic==2.6.0

# Development
pytest==8.0.0
black==24.1.1

httpx==0.27.2
